refactor: tab group types
