+++ projects/ion/src/lib/core/types/index.ts
@@ -20,5 +20,6 @@ export * from './popconfirm';
 export * from './simple-menu';
 export * from './smart-table';
 export * from './status';
+export * from './tab';
 export * from './tab-group';
 export * from './tooltip';
